+++ docente/corsoDiRecuperoVoti.php
@@ -38,15 +38,31 @@ require_once '../common/connect.php';
         <div class="col-md-12">
 <?php
 
+$assenteColor = 'DarkRed';
+$nonRichiestoColor = 'DarkOrange';
+$passatoColor = 'Green';
+$nonPassatoColor = 'FireBrick';
+
 function printableVoto($voto) {
+	global $assenteColor, $nonRichiestoColor, $passatoColor, $nonPassatoColor;
 
 	// rosso se non passato
-	$bgColor = ($voto <= 5) ? 'red' : 'green';
+	if ($voto == 1) {
+		$bgColor = $assenteColor;
+	} elseif ($voto == 2) {
+		$bgColor = $nonRichiestoColor;
+	} else {
+		$bgColor = ($voto <= 5) ? $nonPassatoColor : $passatoColor;
+	}
 
 	if ($voto != 0) {
 		// 1 vuole dire che non c'era
 		if ($voto == 1) {
 			$voto = 'Assente';
+		}
+		// 2 che non ha voluto farlo (eg non ha presentato la richiesta)
+		if ($voto == 2) {
+			$voto = 'Non Richiesto';
 		}
 		$result = '<span class=\'label label-info\' style=\'background-color: '.$bgColor.';\'>'.$voto.'</span>';
 		return $result;
@@ -192,27 +208,37 @@ foreach(dbGetAll($query) as $row) {
 			if ($__config->getVoti_recupero_settembre_aperto()) {
 				$votoSettembre = $studenteRow['studente_per_corso_di_recupero_voto_settembre'];
 				$votoSettembreOptionList = '<select class="votoSettembre selectpicker" data-noneSelectedText="seleziona..." data-width="50%" ><option value="0"></option>';
-
-				// opzione per assente
-				$bgColor = 'red';
+				// opzione per Assente
+				$bgColor = $assenteColor;
 				$votoSettembreOptionList .= '<option value="'.'1'.'" data-content="<span class=\'label label-info\' style=\'background-color: '.$bgColor.';\'>'.'Assente'.'</span>"';
 				if ($votoSettembre == 1) {
 					$votoSettembreOptionList .= ' selected ';
 				}
-				$votoSettembreOptionList .= '>'.'assente'.'</option>';
+				$votoSettembreOptionList .= '>'.'Assente'.'</option>';
 
-				for($i = 4; $i <= 10; $i++) {
-					$bgColor = ($i <= 5) ? 'red' : 'green';
-					// prepara la lista dei possibili voti
-					$votoSettembreOptionList .= '<option value="'.$i.'" data-content="<span class=\'label label-info\'style=\'background-color: '.$bgColor.';\'>'.$i.'</span>"';
+				// opzione per Non Richiesto
+				$bgColor = $nonRichiestoColor;
+				$votoSettembreOptionList .= '<option value="'.'2'.'" data-content="<span class=\'label label-info\' style=\'background-color: '.$bgColor.';\'>'.'Non Richiesto'.'</span>"';
+				if ($votoSettembre == 2) {
+					$votoSettembreOptionList .= ' selected ';
+				}
+				$votoSettembreOptionList .= '>'.'Non Richiesto'.'</option>';
+				
+				// voti da 4 a 10
+				for($i = 4; $i<=10; $i++) {
+					$bgColor = ($i <= 5) ? $nonPassatoColor : $passatoColor;
+					$votoSettembreOptionList .= '<option value="'.$i.'" data-content="<span class=\'label label-info\' style=\'background-color: '.$bgColor.';\'>'.$i.'</span>"';
+
+					// seleziona il voto corrente di settembre se esiste (se e' gia' stato dato)
 					if ($votoSettembre == $i) {
-						// marca selected il voto corrente
 						$votoSettembreOptionList .= ' selected ';
 					}
 					$votoSettembreOptionList .= '>'.$i.'</option>';
 				}
 				$votoSettembreOptionList .= '</select>';
 				$data .= '<td>'.$votoSettembreOptionList.'</td>';
+
+				// la data del voto di settembre
 				$dataSettembre = strftime("%d/%m/%Y", strtotime($studenteRow['studente_per_corso_di_recupero_data_voto_settembre']));
 				if (empty($studenteRow['studente_per_corso_di_recupero_data_voto_settembre'])) {
 					$dataSettembre = date('d/m/Y');